Resolving 747: Reduced and Folded confused. Renamed files and references,  fixed docs.
